uefi: move memory map types from boot.rs to mem/memory_map/mod.rs
This is a preliminary step to further split that module down into multiple
parts.
