+++ uefi/CHANGELOG.md
@@ -24,6 +24,8 @@
 - **Breaking:** `PcrEvent::new_in_buffer` and `PcrEventInputs::new_in_buffer`
   now take an initialized buffer (`[u8`] instead of `[MaybeUninit<u8>]`), and if
   the buffer is too small the required size is returned in the error data.
+- Exports of Memory Map-related types from `uefi::table::boot` are now
+  deprecated. Use `uefi::mem::memory_map` instead.
 
 
 # uefi - 0.29.0 (2024-07-02)
